refactor(fs): move folder browser into the fs namespace as fs.browse
`project.listDirectories` becomes `fs.browse`, so all filesystem access lives
under one `fs` namespace, split by trust boundary in the method signatures:

- fs.browse(path)            — rootless folder picker (names only, no contents)
- fs.readFileString(cwd,...) — cwd-confined read (WorkspaceFSService)
- fs.readDirectory(cwd,...)  — cwd-confined listing

The browse folder-picker policy stays at the router edge (not in the shared
service). `project` keeps only entity ops (list/create). Schemas renamed
ListDirectories* -> Browse*; app import dialog and tests updated.
