+++ packages/contract/src/fs.ts
@@ -1,7 +1,7 @@
 import { oc, type } from "@orpc/contract";
 import { Schema } from "effect";
 
-import { toStandardSchema } from "./domain";
+import { BrowseInputSchema, BrowseResultSchema, toStandardSchema } from "./domain";
 
 // `path` is resolved relative to `cwd` and confined within it on the server.
 const CwdPathInput = Schema.Struct({
@@ -10,10 +10,16 @@ const CwdPathInput = Schema.Struct({
 });
 
 /**
- * Read-only file access, backed by the server's `WorkspaceFSService`. Every read
- * is confined to the caller-supplied `cwd`.
+ * File-system access. `readFileString` / `readDirectory` are confined to the
+ * caller-supplied `cwd` (backed by `WorkspaceFSService`); `browse` is rootless —
+ * a folder picker that may list any directory but returns names only, no
+ * contents.
  */
 export const fsContract = {
   readFileString: oc.input(toStandardSchema(CwdPathInput)).output(type<string>()),
   readDirectory: oc.input(toStandardSchema(CwdPathInput)).output(type<ReadonlyArray<string>>()),
+  /** Browse immediate subdirectories of `path` (default: the home directory). */
+  browse: oc
+    .input(toStandardSchema(BrowseInputSchema))
+    .output(toStandardSchema(BrowseResultSchema)),
 };
